Extract consulting/ bounded context
Move lifecycle domain code (entities, repositories, DTOs, usecases)
from bin/cli/ into the consulting/ package. bin/cli/ files become
re-export shims for backwards compatibility.

The consulting BC owns DecisionEntry, EngagementEntry, all five
repository protocols, eleven consulting DTOs, and eleven usecases.
RegisterTourUseCase, RenderSiteUseCase, and their DTOs remain in
bin/cli/ as they belong to other bounded contexts.

Part of #47.
