+++ tests/conftest.py
@@ -10,7 +10,7 @@ import pytest
 
 from bin.cli.config import Config
 from bin.cli.di import Container
-from bin.cli.entities import DecisionEntry, EngagementEntry
+from consulting.entities import DecisionEntry, EngagementEntry
 from practice.discovery import PipelineStage
 from practice.entities import (
     Confidence,
