+++ tests/test_usecases.py
@@ -17,11 +17,9 @@ from datetime import datetime, timezone
 
 import pytest
 
-from bin.cli.entities import DecisionEntry
+from bin.cli.dtos import RegisterTourRequest
 from bin.cli.wm_types import TourStop
-from practice.exceptions import DuplicateError, InvalidTransitionError, NotFoundError
-from bin.cli.di import Container
-from bin.cli.dtos import (
+from consulting.dtos import (
     AddEngagementEntryRequest,
     GetProjectProgressRequest,
     GetProjectRequest,
@@ -32,9 +30,11 @@ from bin.cli.dtos import (
     RecordDecisionRequest,
     RegisterProjectRequest,
     RegisterResearchTopicRequest,
-    RegisterTourRequest,
     UpdateProjectStatusRequest,
 )
+from consulting.entities import DecisionEntry
+from practice.exceptions import DuplicateError, InvalidTransitionError, NotFoundError
+from bin.cli.di import Container
 
 from .conftest import seed_wardley_mapping
